Complete SGO Refactor
