+++ scripts/bin/account.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# Extract odds data.
+# Extract sports data.
 
 
 #######################################
@@ -30,13 +30,12 @@ info() {
   echo "[$(date +'%Y-%m-%dT%H:%M:%S%z')]: $*"
 }
 
-
 # Begin script execution
-info "Starting odds data extraction..."
+info "Starting account data extraction..."
 
-# Execute the script to fetch odds data
-if node scripts/extract/sports-game-odds/odds.js; then
-  info "Odds data extraction script executed successfully"
+# Execute the script to fetch sports data
+if node scripts/extract/account.js; then
+  info "Account data extraction script executed successfully"
 else
-  err "Failed to execute odds data extraction script"
+  err "Failed to execute account data extraction script"
 fi
