+++ scripts/bin/eventOdds.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 #
-# Extract sport odds data.
+# Extract event odds data.
 
 
 #######################################
@@ -31,11 +31,11 @@ info() {
 }
 
 # Begin script execution
-info "Starting events data extraction..."
+info "Starting event odds data extraction..."
 
 # Execute the script to fetch events data
-if node scripts/extract/the-odds-api/sportOdds.js; then
-  info "Sport odds data extraction script executed successfully"
+if node scripts/extract/eventOdds.js; then
+  info "Events data extraction script executed successfully"
 else
-  err "Failed to execute sport odds data extraction script"
+  err "Failed to execute events data extraction script"
 fi
