+++ scripts/bin/sports.sh
@@ -34,22 +34,8 @@ info() {
 info "Starting sports data extraction..."
 
 # Execute the script to fetch sports data
-if node scripts/extract/the-odds-api/sports.js; then
+if node scripts/extract/sports.js; then
   info "Sports data extraction script executed successfully"
 else
   err "Failed to execute sports data extraction script"
-fi
-
-# Execute the script to fetch events data
-if node scripts/extract/the-odds-api/events.js; then
-  info "Events data extraction script executed successfully"
-else
-  err "Failed to execute events data extraction script"
-fi
-
-# Execute the script to fetch events data
-if node scripts/extract/the-odds-api/eventOdds.js; then
-  info "Event odds data extraction script executed successfully"
-else
-  err "Failed to execute event odds data extraction script"
 fi
