--- scripts/bin/the-odds-api/account.sh
@@ -1,41 +1,0 @@
-#!/bin/bash
-#
-# Extract sports data.
-
-
-#######################################
-# Display error message.
-# Globals:
-#   None
-# Arguments:
-#   None
-# Outputs:
-#   Writes log message to stderr
-#######################################
-err() {
-  echo "[$(date +'%Y-%m-%dT%H:%M:%S%z')]: $*" >&2
-  exit 1
-}
-
-#######################################
-# Display log message.
-# Globals:
-#   None
-# Arguments:
-#   None
-# Outputs:
-#   Writes log message to stdout
-#######################################
-info() {
-  echo "[$(date +'%Y-%m-%dT%H:%M:%S%z')]: $*"
-}
-
-# Begin script execution
-info "Starting account data extraction..."
-
-# Execute the script to fetch sports data
-if node scripts/extract/the-odds-api/account.js; then
-  info "Account data extraction script executed successfully"
-else
-  err "Failed to execute account data extraction script"
-fi
